refactored project structured and remove unnecessary file structure from the project
