--- src/edu/mum/cs/cs525/labs/exercises/project/ui/bank/JDialog_Deposit.java
@@ -1,94 +1,0 @@
-package edu.mum.cs.cs525.labs.exercises.project.ui.bank;
-import java.awt.*;
-import javax.swing.*;
-
-public class JDialog_Deposit extends JDialog
-{
-    
-
-    private BankFrm parentframe;
-    private String accnr;
-    
-	public JDialog_Deposit(BankFrm parent, String aaccnr)
-	{
-		super(parent);
-		parentframe=parent;
-		accnr=aaccnr;
-		// This code is automatically generated by Visual Cafe when you add
-		// components to the visual environment. It instantiates and initializes
-		// the components. To modify the code, only use code syntax that matches
-		// what Visual Cafe can generate, or Visual Cafe may be unable to back
-		// parse your Java file into its visual environment.
-		//{{INIT_CONTROLS
-		setTitle("Deposit");
-		setModal(true);
-		getContentPane().setLayout(null);
-		setSize(268,126);
-		setVisible(false);
-		JLabel1.setText("Acc Nr");
-		getContentPane().add(JLabel1);
-		JLabel1.setForeground(Color.black);
-		JLabel1.setBounds(12,12,48,24);
-		JLabel2.setText("Amount");
-		getContentPane().add(JLabel2);
-		JLabel2.setForeground(Color.black);
-		JLabel2.setBounds(12,48,48,24);
-		JTextField_NAME.setEditable(false);
-		getContentPane().add(JTextField_NAME);
-		JTextField_NAME.setBounds(84,12,144,24);
-		JButton_OK.setText("OK");
-		JButton_OK.setActionCommand("OK");
-		getContentPane().add(JButton_OK);
-		JButton_OK.setBounds(36,84,84,24);
-		JButton_Cancel.setText("Cancel");
-		JButton_Cancel.setActionCommand("Cancel");
-		getContentPane().add(JButton_Cancel);
-		JButton_Cancel.setBounds(156,84,84,24);
-		getContentPane().add(JTextField_Deposit);
-		JTextField_Deposit.setBounds(84,48,144,24);
-		//}}
-	    JTextField_NAME.setText(accnr);
-	    
-		//{{REGISTER_LISTENERS
-		SymAction lSymAction = new SymAction();
-		JButton_OK.addActionListener(lSymAction);
-		JButton_Cancel.addActionListener(lSymAction);
-		//}}
-	}
-
-
-
-	//{{DECLARE_CONTROLS
-	JLabel JLabel1 = new JLabel();
-	JLabel JLabel2 = new JLabel();
-	JTextField JTextField_NAME = new JTextField();
-	JButton JButton_OK = new JButton();
-	JButton JButton_Cancel = new JButton();
-	JTextField JTextField_Deposit = new JTextField();
-	//}}
-
-
-	class SymAction implements java.awt.event.ActionListener
-	{
-		public void actionPerformed(java.awt.event.ActionEvent event)
-		{
-			Object object = event.getSource();
-			if (object == JButton_OK)
-				JButtonOK_actionPerformed(event);
-			else if (object == JButton_Cancel)
-				JButtonCalcel_actionPerformed(event);
-		}
-	}
-
-	void JButtonOK_actionPerformed(java.awt.event.ActionEvent event)
-	{
-        parentframe.amountDeposit=JTextField_Deposit.getText();
-        dispose();
-	}
-
-	void JButtonCalcel_actionPerformed(java.awt.event.ActionEvent event)
-	{
-		dispose();
-	}
-
-}
