--- src/edu/mum/cs/cs525/labs/exercises/project/ui/framework/command/Command.java
@@ -1,6 +1,0 @@
-package edu.mum.cs.cs525.labs.exercises.project.ui.framework.command;
-
-public interface Command {
-    void execute();
-    void undo();
-}
