--- src/edu/mum/cs/cs525/labs/exercises/project/ui/framework/model/Account.java
@@ -1,20 +1,0 @@
-package edu.mum.cs.cs525.labs.exercises.project.ui.framework.model;
-
-import java.time.LocalDate;
-
-public class Account extends BaseEntity{
-    private double currentBalance;
-
-    public Account(Long id, String createdBy, String changedBy, LocalDate createdDate, LocalDate changedDate, double currentBalance) {
-        super(id, createdBy, changedBy, createdDate, changedDate);
-        this.currentBalance = currentBalance;
-    }
-
-    public double getCurrentBalance() {
-        return currentBalance;
-    }
-
-    public void setCurrentBalance(double currentBalance) {
-        this.currentBalance = currentBalance;
-    }
-}
