--- src/edu/mum/cs/cs525/labs/exercises/project/ui/framework/persistence/AccountDao.java
@@ -1,7 +1,0 @@
-package edu.mum.cs.cs525.labs.exercises.project.ui.framework.persistence;
-
-import edu.mum.cs.cs525.labs.exercises.project.ui.framework.model.Account;
-
-public class AccountDao extends Dao<Account> {
-
-}
