+++ src/edu/mum/cs/cs525/project/bank/JDialog_AddPAcc.java
@@ -1,0 +1,194 @@
+package edu.mum.cs.cs525.project.bank;
+/*
+		A basic implementation of the JDialog class.
+*/
+
+import java.awt.*;
+import javax.swing.*;
+
+public class JDialog_AddPAcc extends JDialog
+{
+    private BankFrm parentframe;
+    
+	public JDialog_AddPAcc(BankFrm parent)
+	{
+		super(parent);
+		parentframe=parent;
+		
+		// This code is automatically generated by Visual Cafe when you add
+		// components to the visual environment. It instantiates and initializes
+		// the components. To modify the code, only use code syntax that matches
+		// what Visual Cafe can generate, or Visual Cafe may be unable to back
+		// parse your Java file into its visual environment.
+		//{{INIT_CONTROLS
+		setTitle("Add personal account");
+		setModal(true);
+		getContentPane().setLayout(null);
+		setSize(283,303);
+		setVisible(false);
+		JRadioButton_Chk.setText("Checkings");
+		JRadioButton_Chk.setActionCommand("Checkings");
+		getContentPane().add(JRadioButton_Chk);
+		JRadioButton_Chk.setBounds(36,0,84,24);
+		JRadioButton_Sav.setText("Savings");
+		JRadioButton_Sav.setActionCommand("Savings");
+		getContentPane().add(JRadioButton_Sav);
+		JRadioButton_Sav.setBounds(36,24,84,24);
+		JLabel1.setText("Name");
+		getContentPane().add(JLabel1);
+		JLabel1.setForeground(Color.black);
+		JLabel1.setBounds(12,84,48,24);
+		JLabel2.setText("Street");
+		getContentPane().add(JLabel2);
+		JLabel2.setForeground(Color.black);
+		JLabel2.setBounds(12,108,48,24);
+		JLabel3.setText("City");
+		getContentPane().add(JLabel3);
+		JLabel3.setForeground(Color.black);
+		JLabel3.setBounds(12,132,48,24);
+		JLabel4.setText("State");
+		getContentPane().add(JLabel4);
+		JLabel4.setForeground(Color.black);
+		JLabel4.setBounds(12,156,48,24);
+		JLabel5.setText("Zip");
+		getContentPane().add(JLabel5);
+		JLabel5.setForeground(Color.black);
+		JLabel5.setBounds(12,180,48,24);
+		JLabel6.setText("Birthdate");
+		getContentPane().add(JLabel6);
+		JLabel6.setForeground(Color.black);
+		JLabel6.setBounds(12,204,96,24);
+		JLabel7.setText("Email");
+		getContentPane().add(JLabel7);
+		JLabel7.setForeground(Color.black);
+		JLabel7.setBounds(12,228,48,24);
+		getContentPane().add(JTextField_NAME);
+		JTextField_NAME.setBounds(84,84,156,20);
+		getContentPane().add(JTextField_CT);
+		JTextField_CT.setBounds(84,132,156,20);
+		getContentPane().add(JTextField_ST);
+		JTextField_ST.setBounds(84,156,156,20);
+		getContentPane().add(JTextField_STR);
+		JTextField_STR.setBounds(84,108,156,20);
+		getContentPane().add(JTextField_ZIP);
+		JTextField_ZIP.setBounds(84,180,156,20);
+		getContentPane().add(JTextField_BD);
+		JTextField_BD.setBounds(84,204,156,20);
+		getContentPane().add(JTextField_EM);
+		JTextField_EM.setBounds(84,228,156,20);
+		JButton_OK.setText("OK");
+		JButton_OK.setActionCommand("OK");
+		getContentPane().add(JButton_OK);
+		JButton_OK.setBounds(48,264,84,24);
+		JButton_Cancel.setText("Cancel");
+		JButton_Cancel.setActionCommand("Cancel");
+		getContentPane().add(JButton_Cancel);
+		JButton_Cancel.setBounds(156,264,84,24);
+		getContentPane().add(JTextField_ACNR);
+		JTextField_ACNR.setBounds(84,60,156,20);
+		JLabel8.setText("Acc Nr");
+		getContentPane().add(JLabel8);
+		JLabel8.setForeground(Color.black);
+		JLabel8.setBounds(12,60,48,24);
+		//}}
+	
+		//{{REGISTER_LISTENERS
+		SymMouse aSymMouse = new SymMouse();
+		JRadioButton_Chk.addMouseListener(aSymMouse);
+		JRadioButton_Sav.addMouseListener(aSymMouse);
+		SymAction lSymAction = new SymAction();
+		JButton_OK.addActionListener(lSymAction);
+		JButton_Cancel.addActionListener(lSymAction);
+		//}}
+	}
+
+
+
+
+	//{{DECLARE_CONTROLS
+	JRadioButton JRadioButton_Chk = new JRadioButton();
+	JRadioButton JRadioButton_Sav = new JRadioButton();
+	JLabel JLabel1 = new JLabel();
+	JLabel JLabel2 = new JLabel();
+	JLabel JLabel3 = new JLabel();
+	JLabel JLabel4 = new JLabel();
+	JLabel JLabel5 = new JLabel();
+	JLabel JLabel6 = new JLabel();
+	JLabel JLabel7 = new JLabel();
+	JTextField JTextField_NAME = new JTextField();
+	JTextField JTextField_CT = new JTextField();
+	JTextField JTextField_ST = new JTextField();
+	JTextField JTextField_STR = new JTextField();
+	JTextField JTextField_ZIP = new JTextField();
+	JTextField JTextField_BD = new JTextField();
+	JTextField JTextField_EM = new JTextField();
+	JButton JButton_OK = new JButton();
+	JButton JButton_Cancel = new JButton();
+	JTextField JTextField_ACNR = new JTextField();
+	JLabel JLabel8 = new JLabel();
+	//}}
+
+
+	class SymMouse extends java.awt.event.MouseAdapter
+	{
+		public void mouseClicked(java.awt.event.MouseEvent event)
+		{
+			Object object = event.getSource();
+			if (object == JRadioButton_Chk)
+				JRadioButtonChk_mouseClicked(event);
+			else if (object == JRadioButton_Sav)
+				JRadioButtonSav_mouseClicked(event);
+		}
+	}
+
+	void JRadioButtonChk_mouseClicked(java.awt.event.MouseEvent event)
+	{
+		//When Checking radio is clicked make this radio on 
+		//and make Saving account radio off
+		JRadioButton_Chk.setSelected(true);
+		JRadioButton_Sav.setSelected(false);
+	}
+
+	void JRadioButtonSav_mouseClicked(java.awt.event.MouseEvent event)
+	{
+		//When Saving radio is clicked make this radio on 
+		//and make Checking account radio off
+		JRadioButton_Chk.setSelected(false);
+		JRadioButton_Sav.setSelected(true);
+	 
+	}
+
+	class SymAction implements java.awt.event.ActionListener
+	{
+		public void actionPerformed(java.awt.event.ActionEvent event)
+		{
+			Object object = event.getSource();
+			if (object == JButton_OK)
+				JButtonOK_actionPerformed(event);
+			else if (object == JButton_Cancel)
+				JButtonCalcel_actionPerformed(event);
+		}
+	}
+
+	void JButtonOK_actionPerformed(java.awt.event.ActionEvent event)
+	{
+       parentframe.accountnr=JTextField_ACNR.getText();
+       parentframe.clientName=JTextField_NAME.getText();
+       parentframe.street=JTextField_STR.getText();
+       parentframe.city=JTextField_CT.getText();
+       parentframe.zip=JTextField_ZIP.getText();
+       parentframe.state=JTextField_ST.getText();
+       if (JRadioButton_Chk.isSelected())
+           parentframe.accountType="Ch";
+           else
+           parentframe.accountType="S";
+	   parentframe.newaccount=true;
+       dispose();
+	}
+
+	void JButtonCalcel_actionPerformed(java.awt.event.ActionEvent event)
+	{
+    //make this frame invisible if Cancel button is clicked
+        dispose();
+	}
+}
