+++ src/edu/mum/cs/cs525/project/framework/controller/command/Command.java
@@ -1,0 +1,6 @@
+package edu.mum.cs.cs525.project.framework.controller.command;
+
+public interface Command {
+    void execute();
+    void undo();
+}
