+++ src/edu/mum/cs/cs525/project/framework/persistence/AccountDao.java
@@ -1,0 +1,7 @@
+package edu.mum.cs.cs525.project.framework.persistence;
+
+import edu.mum.cs.cs525.project.framework.model.Account;
+
+public class AccountDao extends Dao<Account> {
+
+}
